refactor(make calls belongs to lead contact and deal)

+++ database/migrations/2025_04_08_131503_create_crm_lead_conversations_table.php
@@ -24,7 +24,7 @@ return new class extends Migration {
            $table->foreignIdFor(\App\Models\Deal::class)
                ->constrained('crm_deals');
 
-            $table->unsignedBigInteger('converted_by');
+            $table->unsignedInteger('converted_by');
             $table->foreign('converted_by')
                 ->references('id')
                 ->on('users');
